move Type files into dir
