+++ packages/quicktype-core/src/Type/TypeUtils.ts
@@ -1,14 +1,14 @@
 import { EqualityMap, iterableFirst, setFilter, setSortBy, setUnion } from "collection-utils";
 
 // eslint-disable-next-line import/no-cycle
-import { type StringTypes, stringTypesTypeAttributeKind } from "./attributes/StringTypes";
+import { type StringTypes, stringTypesTypeAttributeKind } from "../attributes/StringTypes";
 import {
     type CombinationKind,
     type TypeAttributes,
     combineTypeAttributes,
     emptyTypeAttributes
-} from "./attributes/TypeAttributes";
-import { assert, assertNever, defined, panic } from "./support/Support";
+} from "../attributes/TypeAttributes";
+import { assert, assertNever, defined, panic } from "../support/Support";
 import {
     ArrayType,
     type ClassProperty,
@@ -21,7 +21,7 @@ import {
     type Type,
     UnionType
 } from "./Type";
-import { isPrimitiveStringTypeKind } from "./Type/TransformedStringType";
+import { isPrimitiveStringTypeKind } from "./TransformedStringType";
 
 export function assertIsObject(t: Type): ObjectType {
     if (t instanceof ObjectType) {
