+++ packages/quicktype-core/src/Type/index.ts
@@ -1,5 +1,8 @@
 // eslint-disable-next-line import/no-cycle
 export * from "./Type";
+export * from "./TypeBuilder";
+export * from "./TypeGraph";
+export * from "./TypeUtils";
 
 // TODO: update this to type * post TS upgrade
 export * from "./TransformedStringType";
