+++ packages/quicktype-core/src/index.ts
@@ -74,12 +74,12 @@ export { readableFromFileOrURL, readFromFileOrURL } from "./input/io/NodeIO";
 export { FetchingJSONSchemaStore } from "./input/FetchingJSONSchemaStore";
 export { JSONSchemaStore, type JSONSchema } from "./input/JSONSchemaStore";
 export { sourcesFromPostmanCollection } from "./input/PostmanCollection";
-export { TypeBuilder, type StringTypeMapping } from "./TypeBuilder";
-export { type TypeRef, derefTypeRef } from "./TypeGraph";
+export { TypeBuilder, type StringTypeMapping } from "./Type/TypeBuilder";
+export { type TypeRef, derefTypeRef } from "./Type/TypeGraph";
 export { TypeAttributeKind, type TypeAttributes, emptyTypeAttributes } from "./attributes/TypeAttributes";
 export { TypeNames, makeNamesTypeAttributes, namesTypeAttributeKind } from "./attributes/TypeNames";
 export { StringTypes } from "./attributes/StringTypes";
-export { removeNullFromUnion, matchType, nullableFromUnion } from "./TypeUtils";
+export { removeNullFromUnion, matchType, nullableFromUnion } from "./Type/TypeUtils";
 export { ConvenienceRenderer } from "./ConvenienceRenderer";
 export { uriTypeAttributeKind } from "./attributes/URIAttributes";
